refactor: move config file resolvers into local_config
