+++ linkup-cli/src/start.rs
@@ -6,6 +6,7 @@ use std::{
     path::{Path, PathBuf},
 };
 
+use crate::local_config::{config_path, get_config};
 use crate::{
     background_booting::boot_background_services,
     linkup_file_path,
@@ -53,54 +54,6 @@ pub fn start(config_arg: Option<String>) -> Result<(), CliError> {
     check_local_not_started()?;
 
     Ok(())
-}
-
-// TODO(augustoccesar)[2023-09-22]: Move this into a shared file. Maybe local_config?
-pub fn config_path(config_arg: &Option<String>) -> Result<String, CliError> {
-    match config_arg {
-        Some(path) => {
-            let absolute_path = fs::canonicalize(path)
-                .map_err(|_| CliError::NoConfig("Unable to resolve absolute path".to_string()))?;
-            Ok(absolute_path.to_string_lossy().into_owned())
-        }
-        None => match env::var(LINKUP_CONFIG_ENV) {
-            Ok(val) => {
-                let absolute_path = fs::canonicalize(val).map_err(|_| {
-                    CliError::NoConfig("Unable to resolve absolute path".to_string())
-                })?;
-                Ok(absolute_path.to_string_lossy().into_owned())
-            }
-            Err(_) => Err(CliError::NoConfig(
-                "No config argument provided and LINKUP_CONFIG environment variable not set"
-                    .to_string(),
-            )),
-        },
-    }
-}
-
-// TODO(augustoccesar)[2023-09-22]: Move this into a shared file. Maybe local_config?
-pub fn get_config(config_path: &str) -> Result<YamlLocalConfig, CliError> {
-    let content = match fs::read_to_string(config_path) {
-        Ok(content) => content,
-        Err(_) => {
-            return Err(CliError::BadConfig(format!(
-                "Failed to read the config file at {}",
-                config_path
-            )))
-        }
-    };
-
-    let yaml_config: YamlLocalConfig = match serde_yaml::from_str(&content) {
-        Ok(config) => config,
-        Err(_) => {
-            return Err(CliError::BadConfig(format!(
-                "Failed to deserialize the config file at {}",
-                config_path
-            )))
-        }
-    };
-
-    Ok(yaml_config)
 }
 
 pub fn get_state() -> Result<LocalState, CliError> {
